rename dylib
